requirements.txt
PyYAML==6.0.2
pydantic==2.11.3
requests==2.32.3
tenacity==9.0.0
pdfplumber==0.11.5
pypdf==5.4.0
python-docx==1.1.2
reportlab==4.4.0
python-pptx==1.0.2
jsonschema==4.23.0
rapidfuzz==3.13.0
Pillow==11.1.0
